+++ packages/extension/src/progressView/frontend/components/WorkflowStreamContent.ts
@@ -19,9 +19,7 @@ import './LogList';
 import './UsagePanel';
 import './FileList';
 import './WorkflowToolUseFollowupSection';
-import './BackgroundTasksPanel';
 import './RequestPanels';
-import './WorkflowHintBanner';
 
 @customElement('workflow-stream-content')
 export class WorkflowStreamContent extends BaseStreamContent {
@@ -49,15 +47,9 @@ export class WorkflowStreamContent extends BaseStreamContent {
 
       <div class="conversation-content">
         <div class="conversation-column conversation-prelude">
-          <workflow-hint-banner></workflow-hint-banner>
-
           <request-panels
             .permissions=${this.filteredPermissions}
           ></request-panels>
-
-          <background-tasks-panel
-            .subagents=${state.subagents}
-          ></background-tasks-panel>
         </div>
 
         <div class="conversation-log"><log-list></log-list></div>